requirements.txt
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.2
django-allauth==0.39.1
django-environ==0.4.5
django-rest-auth==0.9.5
djangorestframework==3.9.4
djangorestframework-simplejwt==4.3.0
idna==2.8
oauthlib==3.0.1
Pillow==6.0.0
psycopg2==2.8.3
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
